edgeaudit/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
numpy
pandas
scikit-learn
torch
xgboost
scipy
snowflake-connector-python
google-generativeai
httpx
yfinance
pytest
